upload homework

+++ TP_JpnPark/TP_JpnPark-ejb/src/main/java/tp/jpnpark/services/ParkService.java
@@ -11,6 +11,7 @@ import tp.jpnpark.entities.Machine;
 import tp.jpnpark.entities.Park;
 import tp.jpnpark.entities.Visitor;
 import tp.jpnpark.enums.StateOfTheVisitor;
+import tp.jpnpark.exceptions.InvalidValues;
 import tp.jpnpark.facade.EntityFacade;
 
 /**
@@ -30,8 +31,12 @@ public class ParkService {
     private VisitorService visitorService;
 
     public Park create(Park park, long addressId) {
-        park.setAdress(entityManager.find(Address.class, addressId));
-        return entityManager.create(park);
+        if (entityManager.find(Address.class, addressId) != null) {
+            park.setAdress(entityManager.find(Address.class, addressId));
+            return entityManager.create(park);
+        }
+        throw new InvalidValues("There is no address with this ID");
+
     }
 
     public Park find(long parkId) {
@@ -39,26 +44,20 @@ public class ParkService {
     }
 
     public Park update(Park park, long parkId) {
-        Park tempPark = entityManager.find(Park.class, parkId); // eddigi adatai az adott parknak
+        Park tempPark = entityManager.find(Park.class, parkId);
 
         park.setAdress(tempPark.getAdress());
         park.setId(tempPark.getId());
         park.setMachines(tempPark.getMachines());
-        park.setArea(tempPark.getArea() + park.getArea()); // csak bővíteni lehet, csükkenteni nem lehet egy már létező park méretét. 
+        park.setArea(tempPark.getArea() + park.getArea());
 
-        for (Visitor visitor : entityManager.findAll(Visitor.class)) {// átállítani minden visitor nál a parkot
+        for (Visitor visitor : entityManager.findAll(Visitor.class)) {
             if (visitor.getPark().getId().equals(park.getId())) {
                 visitor.setPark(park);
                 entityManager.update(visitor);
             }
         }
 
-        for (Machine machine : entityManager.findAll(Machine.class)) {// átállítani minden machine nél a parkot
-            if (machine.getPark().getId().equals(park.getId())) {
-                machine.setPark(park);
-                entityManager.update(machine);
-            }
-        }
         for (GuestBook gb : entityManager.findAll(GuestBook.class)) {// átállítani minden guestBook nál a parkot
             if (gb.getPark().getId().equals(park.getId())) {
                 gb.setPark(park);
@@ -68,22 +67,29 @@ public class ParkService {
         return entityManager.update(park);
     }
 
-    public void Delete(long parkId) {
-        deleteMachinesFromPark(parkId);
+    public void delete(long parkId) {
+        checkPark(parkId);
+        Park tempPark = entityManager.find(Park.class, parkId);
+        if (tempPark.getMachines() != null) {
+            deleteMachinesFromPark(parkId);
+        }
         deleteVisitorsFromPark(parkId);
         deleteGuestBookOfThePark(parkId);
-        entityManager.delete(parkId);
+        entityManager.delete(tempPark);
     }
 
     public void deleteMachinesFromPark(long parkId) {
         Park tempPark = entityManager.find(Park.class, parkId);
         List<Machine> machines = tempPark.getMachines();
 
-        for (Machine machine : machines) {
-            machinService.deleteVisitorsFromMachine(machine);
-            machine.setPark(null);
-            entityManager.update(machine);
+        if (tempPark.getMachines() != null) {
+            for (Machine machine : machines) {
+                machinService.deleteVisitorsFromMachine(machine);
+                entityManager.update(machine);
+                machine.setParkID(null);
+            }
         }
+
         tempPark.getMachines().clear();
         entityManager.update(tempPark);
     }
@@ -92,7 +98,8 @@ public class ParkService {
         Park tempPark = entityManager.find(Park.class, parkId);
         Machine machine = entityManager.find(Machine.class, machineId);
         machinService.deleteVisitorsFromMachine(machine);
-        machine.setPark(null);
+        tempPark.getMachines().remove(machine);
+        machine.setParkID(null);
         entityManager.update(machine);
         entityManager.update(tempPark);
     }
@@ -121,7 +128,7 @@ public class ParkService {
         if (entityManager.find(Park.class, parkId) != null) {
             return true;
         }
-        throw new RuntimeException("there is no park with this ID");
+        throw new InvalidValues("there is no park with this ID");
     }
 
     public void addMachine(long parkId, long machineId) {
@@ -132,16 +139,19 @@ public class ParkService {
         Machine machine = entityManager.find(Machine.class, machineId);
 
         if (park.getMoney() < machine.getPrice()) {
-            throw new RuntimeException("the park has not enough money to byu this machine");
+            throw new InvalidValues("the park has not enough money to byu this machine");
         }
 
         if (park.getArea() < machine.getSizeOfMachine()) {
-            throw new RuntimeException("there is not enough space to get this machine");
+            throw new InvalidValues("there is not enough space to get this machine");
         }
 
         park.getMachines().add(machine);
         park.setMoney(park.getMoney() - machine.getPrice());
+        park.setArea(park.getArea() - machine.getSizeOfMachine());
+        machine.setParkID(park.getId());
         entityManager.update(park);
+        entityManager.update(machine);
     }
 
     public void addVisitor(long visitorId, long parkId) {
@@ -152,7 +162,7 @@ public class ParkService {
         Visitor visitor = entityManager.find(Visitor.class, visitorId);
 
         if (visitor.getMoney() < park.getTicketPrice()) {
-            throw new RuntimeException("the visitor has no money to buy a ticket");
+            throw new InvalidValues("the visitor has no money to buy a ticket");
         }
 
         visitor.setIsActive(true);
@@ -170,9 +180,7 @@ public class ParkService {
         Park park = entityManager.find(Park.class, parkId);
         Visitor visitor = entityManager.find(Visitor.class, visitorId);
 
-        if (!visitor.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("the visitor is not in this park");
-        }
+        visitorNotInPark(visitor, park);
 
         visitor.setIsActive(false);
         visitor.setPark(null);
@@ -187,18 +195,15 @@ public class ParkService {
         Park park = entityManager.find(Park.class, parkId);
         Visitor visitor = entityManager.find(Visitor.class, visitorId);
         Machine machine = entityManager.find(Machine.class, machineId);
-
-        if (!visitor.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("the visitor is not in this park");
-        }
-        if (!machine.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("the machine is not in this park");
+        visitorNotInPark(visitor, park);
+        if (!park.getMachines().contains(machine)) {
+            throw new InvalidValues("the machine is not in this park");
         }
         if (visitor.getState().equals(StateOfTheVisitor.ON_MACHINE)) {
-            throw new RuntimeException("the visitor is already on a machine");
+            throw new InvalidValues("the visitor is already on a machine");
         }
         if (machine.getCapacity() == machine.getVisitorsOnMachine()) {
-            throw new RuntimeException("there is no free spaces on the machine for the visitor to get in");
+            throw new InvalidValues("there is no free spaces on the machine for the visitor to get in");
         }
 
         machine.setVisitorsOnMachine(machine.getVisitorsOnMachine() + 1);
@@ -218,17 +223,15 @@ public class ParkService {
         Visitor visitor = entityManager.find(Visitor.class, visitorId);
         Machine machine = entityManager.find(Machine.class, machineId);
 
-        if (!visitor.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("the visitor is not in this park");
-        }
-        if (!machine.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("the machine is not in this park");
+        visitorNotInPark(visitor, park);
+        if (!park.getMachines().contains(machine)) {
+            throw new InvalidValues("the machine is not in this park");
         }
         if (visitor.getState().equals(StateOfTheVisitor.REST)) {
-            throw new RuntimeException("the visitor is not on a machine");
+            throw new InvalidValues("the visitor is not on a machine");
         }
         if (!visitor.getMachine().getId().equals(machine.getId())) {
-            throw new RuntimeException("the visitor is not on this machine");
+            throw new InvalidValues("the visitor is not on this machine");
         }
 
         visitor.setState(StateOfTheVisitor.REST);
@@ -238,25 +241,23 @@ public class ParkService {
         entityManager.update(machine);
     }
 
-    public void LogIntoGuestBook(long parkId, long visitorId, long guestBookId, String mesage) {
+    public void logIntoGuestBook(long parkId, long visitorId, long guestBookId, String mesage) {
         visitorService.checkVisitor(visitorId);
         checkPark(parkId);
 
         //check if guestbook is exists
         if (entityManager.find(GuestBook.class, guestBookId) == null) {
-            throw new RuntimeException("there is no guestbook with this id");
+            throw new InvalidValues("there is no guestbook with this id");
         }
 
         GuestBook gb = entityManager.find(GuestBook.class, guestBookId);
         Park park = entityManager.find(Park.class, parkId);
         Visitor visitor = entityManager.find(Visitor.class, visitorId);
 
-        if (!visitor.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("the visitor is not in this park");
-        }
+        visitorNotInPark(visitor, park);
 
         if (!gb.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("this guestbook is not belong to this park");
+            throw new InvalidValues("this guestbook is not belong to this park");
         }
 
         gb.getLogs().put(LocalDate.now(), mesage);
@@ -266,7 +267,7 @@ public class ParkService {
 
     public String getGuestBookLogs(long parkId, long gbId) {
         if (entityManager.find(GuestBook.class, gbId) == null) {
-            throw new RuntimeException("there is no guestbook with this id");
+            throw new InvalidValues("there is no guestbook with this id");
         }
         checkPark(parkId);
 
@@ -274,9 +275,16 @@ public class ParkService {
         Park park = entityManager.find(Park.class, parkId);
 
         if (!gb.getPark().getId().equals(park.getId())) {
-            throw new RuntimeException("this guestbook is not belong to this park");
+            throw new InvalidValues("this guestbook is not belong to this park");
         }
 
         return gb.getLogs().toString();
     }
+
+    public void visitorNotInPark(Visitor visitor, Park park) {
+        if (!visitor.getPark().getId().equals(park.getId())) {
+            throw new InvalidValues("the visitor is not in this park");
+        }
+    }
+
 }
